login, boe user, boe, open ai api.

+++ src/main/java/com/paellasoft/CRUD/repository/IUserRepository.java
@@ -3,7 +3,7 @@ package com.paellasoft.CRUD.repository;
 import com.paellasoft.CRUD.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepository extends JpaRepository<User, Long> {
+public interface IUserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 
 }
